requirements.txt
cffi==1.2.1
configparser==3.3.0.post2
cryptography==1.0.2
DateTime==4.0.1
enum34==1.0.4
google-api-python-client==1.4.2
gspread==0.2.5
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
oauth2client==1.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyGithub==1.25.2
pyOpenSSL==0.15.1
pytz==2015.6
rsa==3.2
simplejson==3.8.0
six==1.10.0
uritemplate==0.6
wheel==0.24.0
zope.interface==4.1.3
